flamingo/requirements.txt
asgiref==3.5.0
Django==3.2.11
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
